вынес все методы валидации данных в отдельный файл
